The stub reconstruction lands with its own cast debt paid: 360 -> 333
reinterpret_casts, and the OPEN worklist is at 0 for the first time. All 27 were
folded by fixing the mis-modeled declaration - none parked, none converted to
another cast form.

Reconstructed: CreateActionArea, DrawPageDebugText, EditDwRectsDlgProc,
CPlay::DrawCustomLevelBanner and CNetSession::DispatchMsg all 0-3% -> 100 EXACT;
CWarlord::SerializeMove 0.19 -> 99.52; CHashElement::Prev 6.90 -> 99.66;
DrawDebugStatsFull 0.42 -> 99.10; CState::DrawScreenTextImage 1.83 -> 97.15;
CMulti::HandleControlMsg 0 -> 96.66.

The four cast folds, each of which killed a stated reason:

  * WrapCoord takes LONG*, not i32*. DrawObjectCounts @0x15a712 hands it
    `lea ecx,[esp+0x20]` / `lea eax,[esp+0x24]` - the .right/.bottom of the same
    4-dword rect it later passes to DrawTextA, which makes it a genuine Win32
    RECT. -18. (My brief claimed every WrapCoord call site was in WwdObjMgr.cpp;
    there are ten more, 8 of them in CImage.cpp. They only take the address of a
    coordinate local, so they follow the declaration and nothing moves.)
  * The map lookups came out OUTRIGHT, no CTypedPtrMap needed. The prior
    "not foldable" verdict assumed the signature was `void*&`; afxcoll.h
    actually declares `BOOL Lookup(LPCTSTR, CObject*&)`, so a CObject* local
    binds with NO cast and the derived type is recovered by a real static_cast.
    -5, all four DispatchKeyed* and InstallTree still 100.00 EXACT.
  * CSBI_WellGoo::Setup's 9th arg. Its note claimed retail's mangling proved
    `int` - but that mangled name is OUR OWN label, so the argument cited itself.
    The real proof is the sibling slot-11 CSBI_Image::SetupImage, handed a
    DATA-relocated string literal (push 0x614168 = "GAME_STATUSBAR_TABZ_
    STATZTAB"). i32 a9 -> const char* key across the family. -3.
  * Play.cpp -1, and this one cost 4.4 points deliberately. The 99.10% was
    propped up by ALIASING a char[0x40] sprintf buffer as a RECT to force the
    stack colouring. The bytes show why retail gets that colouring for free: the
    same slot IS a sprintf destination earlier in the body, so it is
    address-exposed, and no source spelling expresses that. Honest `RECT lr;`,
    94.72%, @early-stop rewritten with the mechanism so nobody re-puns it.

Also corrected the rva-span pattern doc, which read as "always grow the span" -
the tileswitchlogic counter-case (three functions made WORSE by the same move)
was missing.

NOTE, flagged by the lane: m_<hex> fields rose 7483 -> 7581 from the 14 new
bodies and the build blessed it, because that metric is not in _RATCHET. Real
naming debt; ratcheting it is the next commit.

gruntz build --full: exit 0. 3,161 exact (was 3,148). Fuzzy 80.07%, MAX 80.16.

+++ src/Gruntz/TriggerMgr.cpp
@@ -839,8 +839,8 @@ i32 CTriggerMgr::ReinitGroup(i32 col, i32 row) {
     // the main plane's coord wrap (thunk 0x295a -> ?WrapCoord@CDDrawWorkerHost@@ @0xa000;
     // receiver is level->m_mainPlane)
     CGameLevel* plane = g_gameReg->m_world->m_level;
-    i32 outR = col;
-    i32 outC = row;
+    LONG outR = col;
+    LONG outC = row;
     plane->m_mainPlane->WrapCoord(&outR, &outC);
     CStatusBarMgr* sbi = lvl->m_guts;
     if (sbi->m_hlBusy == 0) {
